refactor(web/mcp): extract TryItForm with onRun callback
